Create Commit #2
In this code, using PyCharm, I have made edits to the original code. Here, I have turned the game into a 'Virtual Pet' experience. The goal of this Virtual Pet, is to grow it based on completing real-life activties. The program is intended to motivate the average person to complete tasks and chores in daily life, and use the Virtual Pet as a reward system. 

Changes: Ability to name pet, an 'age' factor that increases with completed tasks. I also toyed with a 'feed' option but decided not to follow through. This specific code will not run as it has some errors.